+++ pywatts/summaries/metric_base.py
@@ -1,15 +1,17 @@
 import logging
 from abc import ABC, abstractmethod
-from typing import Callable, Tuple, Dict, Optional
+from typing import Callable, Tuple, Dict, Optional, List
 
 import cloudpickle
 import numpy as np
+import pandas as pd
 import xarray as xr
 
 from pywatts.core.base_summary import BaseSummary
-from pywatts.core.summary_object import SummaryObject, SummaryObjectList
 from pywatts.core.exceptions import InputNotAvailable
+from pywatts.core.exceptions.invalid_input_exception import InvalidInputException
 from pywatts.core.filemanager import FileManager
+from pywatts.core.summary_object import SummaryObjectList
 
 logger = logging.getLogger(__name__)
 
@@ -28,10 +30,15 @@ class MetricBase(BaseSummary, ABC):
     def __init__(self,
                  name: str = None,
                  filter_method: Callable[[np.ndarray, np.ndarray], Tuple[np.ndarray, np.ndarray]] = None,
-                 offset: int = 0):
+                 offset: int = 0,
+                 cuts: List[Tuple[pd.Timestamp, pd.Timestamp]] = None):
         super().__init__(name if name is not None else self.__class__.__name__)
         self.offset = offset
         self.filter_method = filter_method
+        if cuts is None:
+            self.cuts = []
+        else:
+            self.cuts = cuts
 
     def get_params(self) -> Dict[str, object]:
         """
@@ -40,17 +47,22 @@ class MetricBase(BaseSummary, ABC):
         :return: Parameters set for the Metric
         :rtype: Dict[str, object]
         """
-        return {"offset": self.offset}
+        return {"offset": self.offset,
+                "cuts": self.cuts}
 
-    def set_params(self, offset: Optional[int] = None):
+    def set_params(self, offset: Optional[int] = None, cuts=Optional[List[Tuple[pd.Timestamp, pd.Timestamp]]]):
         """
         Set parameters of the Metric.
 
         :param offset: Offset, which determines the number of ignored values in the beginning for calculating the Metric.
         :type offset: int
+        :param cuts: The cutouts on which the metric should be additionally calculated.
+        :type cuts: List[Tuple[pd.Timestamp, pd.Timestamp]]
         """
         if offset:
             self.offset = offset
+        if cuts is not None:
+            self.cuts = cuts
 
     def transform(self, file_manager: FileManager, y: xr.DataArray, **kwargs: xr.DataArray) -> SummaryObjectList:
         """
@@ -74,15 +86,34 @@ class MetricBase(BaseSummary, ABC):
             logger.error(error_message)
             raise InputNotAvailable(error_message)
 
+        y_ = y[self.offset:]
+        kwargs_ = {key: value[self.offset:] for key, value in kwargs.items()}
+        self._transform({key: value for key, value in kwargs_.items()}, ": complete", summary, y_.values)
+        for start, end in self.cuts:
+            _kwargs = {key: value.loc[start:end] for key, value in kwargs_.items()}
+            if not len(y_.loc[start:end].values) == 0:
+                self._transform(_kwargs, f": Cut from {start} to {end}", summary, y_.loc[start:end].values)
+
+        return summary
+
+    def _transform(self, kwargs, suffix, summary, t):
         for key, y_hat in kwargs.items():
             p = y_hat.values
+            if p.shape != t.shape:
+                try:
+                    p = p.reshape(t.shape)
+                except ValueError:
+                    raise InvalidInputException(
+                        f"The prediction {key} does not match to the shape of the ground truth y in the instance "
+                        f"{self.name} of class {self.__class__.__name__}.")
+                self.logger.info(f"Reshaped prediction {key} in {self.name}")
             if self.filter_method:
                 p_, t_ = self.filter_method(p, t)
                 mae = self._apply_metric(p_, t_)
 
             else:
                 mae = self._apply_metric(p, t)
-            summary.set_kv(key, mae)
+            summary.set_kv(key + suffix, mae)
         return summary
 
     def save(self, fm: FileManager) -> Dict: