+++ CHANGELOG.md
@@ -3,9 +3,17 @@
 ## Unreleased
 
 ### Added
-* Add average module for pre-processing
+* Add average module for pre-processing ([#166](https://github.com/KIT-IAI/pyWATTS/issues/166))
+* Metric can be calculated on cutouts ([#149](https://github.com/KIT-IAI/pyWATTS/issues/149))
+* Add custom scaler module for pre-processing ([#167](https://github.com/KIT-IAI/pyWATTS/issues/167))
+* Add a MASE Summary ([#148](https://github.com/KIT-IAI/pyWATTS/issues/148))
+
 
 ### Changed
+
+### Fixed
+* Try to handle non matching shapes in metrics. If not possible raise an exception.  ([#154](https://github.com/KIT-IAI/pyWATTS/issues/154))
+
 
 ### Deprecated